requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
appnope==0.1.2
APScheduler==3.6.3
asttokens==2.0.5
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bleach==5.0.1
cachetools==4.2.2
certifi==2022.12.7
charset-normalizer==2.1.1
contourpy==1.0.6
cycler==0.11.0
dataframe-image==0.1.3
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.16.2
fonttools==4.38.0
frozenlist==1.3.3
google-auth==2.15.0
google-auth-oauthlib==0.8.0
gspread==5.7.2
gspread-dataframe==3.3.0
idna==3.4
importlib-metadata==5.2.0
importlib-resources==5.10.1
ipykernel==6.15.2
ipython==8.7.0
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.17.3
jupyter_core==5.1.1
jupyterlab-pygments==0.2.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==2.0.4
multidict==6.0.4
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.1
nest-asyncio==1.5.6
numpy==1.24.1
oauthlib==3.2.2
openai==0.27.7
packaging==22.0
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pip==22.3.1
pkgutil_resolve_name==1.3.10
platformdirs==2.6.0
prompt-toolkit==3.0.20
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-dotenv==0.20.0
python-telegram-bot==13.11
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
pyzmq==24.0.1
requests>=2.31.0
requests-oauthlib==1.3.1
rsa==4.9
setuptools==65.5.0
six==1.16.0
soupsieve==2.3.2.post1
stack-data==0.2.0
tinycss2==1.2.1
tornado>=6.3.2
traitlets==5.8.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
wcwidth==0.2.5
webencodings==0.5.1
wheel==0.37.1
yarl==1.8.2
zipp==3.11.0
